requirements.txt
asgiref==3.8.1
Django==5.0.4
django-filter==24.2
django-templates==0.0.13
django-templatetags==1.1
django-tree-queries==0.19.0
psycopg==3.1.18
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
